Feat : 스타벅스 카드 등록, 조회
- (관리자)스타벅스 카드 발급 번호+핀번호 저장
- 스타벅스 카드 등록
- 등록 된 스타벅스 카드 리스트 조회

Resolves: #7

+++ src/main/java/com/example/shinsekai/card/entity/StarbucksCard.java
@@ -31,7 +31,7 @@ public class StarbucksCard extends BaseEntity {
     @Column(nullable = false, length = 100)
     private String cardImageUrl;//섬네일 이미지
     private String cardDescription;
-    private int remainAmount;
+    private double remainAmount;
 
     @Builder
     public StarbucksCard(
@@ -42,7 +42,7 @@ public class StarbucksCard extends BaseEntity {
             String cardNumber,
             String cardImageUrl,
             String cardDescription,
-            int remainAmount) {
+            double remainAmount) {
 
         this.id = id;
         this.cardUuid = cardUuid;
